refactor(backend): split providers/llm.py into the providers/llm package
Pure mechanical move of the 2688-line monolith into focused submodules:
client (AsyncOpenAI plumbing, provider/tier resolution, _complete_json
ladder, retries, JSON helpers), planner (plan_page + prompt polish),
click (click/stroke resolution + candidate precompute), extraction
(entity extraction), world (neighbors, NL geo-edits, world planning).

__init__ re-exports the entire previous surface, so providers.llm
resolves identically. Submodules call the test-monkeypatched seams
(_client, _complete_json, _safe_log) through the package namespace so
setattr on providers.llm keeps intercepting internal calls, exactly as
on the monolith. No behavior change; gates: ruff, mypy, pytest (811
passed, 2 pre-existing skips).
